+++ tools/utils/images/to_rgba.go
@@ -8,9 +8,12 @@ import (
 	"math"
 
 	"github.com/kovidgoyal/imaging"
+	"github.com/kovidgoyal/kitty/tools/tty"
 )
 
 var _ = fmt.Print
+var debugprintln = tty.DebugPrintln
+var _ = debugprintln
 
 func (self *Context) run_paste(src imaging.Scanner, background image.Image, pos image.Point, postprocess func([]byte)) {
 	pos = pos.Sub(background.Bounds().Min)
@@ -24,15 +27,16 @@ func (self *Context) run_paste(src imaging.Scanner, background image.Image, pos 
 	var pix []uint8
 	switch v := background.(type) {
 	case *image.NRGBA:
-		i := background.(*image.NRGBA)
-		stride = i.Stride
-		pix = i.Pix
+		stride = v.Stride
+		pix = v.Pix
 	case *imaging.NRGB:
-		i := background.(*imaging.NRGB)
-		stride = i.Stride
-		pix = i.Pix
+		stride = v.Stride
+		pix = v.Pix
 	default:
 		panic(fmt.Sprintf("Unsupported image type: %v", v))
+	}
+	if len(pix) < background.Bounds().Dy()*stride {
+		panic(fmt.Sprintf("background image has insufficient pixel data. Bounds: %v Stride: %d Data len: %d", background.Bounds(), stride, len(pix)))
 	}
 	if err := self.SafeParallel(interRect.Min.Y, interRect.Max.Y, func(ys <-chan int) {
 		for y := range ys {